Adicionando confirmação de delete

+++ src/app/task/task.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input } from '@angular/core';
 import WebService from '../../webservice';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-task',
@@ -16,6 +17,24 @@ export class TaskComponent {
   }
 
   excluirTarefa(id: number){
-    this.WebService.excluirTarefa(id);
+    Swal.fire({
+      animation: true,
+      background: '#1A1A1A',
+      color: 'white',
+      title: "Deletar Tarefa",
+      text: "Deseja realmente deletar essa tarefa?",
+      confirmButtonText: "Sim",
+      confirmButtonColor: '#1E6F9F',
+      confirmButtonAriaLabel: "Apagar Tarefa",
+      denyButtonText: "Cancelar",
+      denyButtonColor: '#5E60CE',
+      denyButtonAriaLabel: "Cancelar a ação",
+      showCloseButton: true,
+      showDenyButton: true,
+    }).then((result)=>{
+      if(result.isConfirmed){
+        this.WebService.excluirTarefa(id);
+      }
+    })
   }
 }
